% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_phom.R
\name{calcPhom}
\alias{calcPhom}
\title{Calculate persistent diagram of point cloud \code{X}}
\usage{
calcPhom(X, maxdimension, maxscale, plot = T, ret = T)
}
\arguments{
\item{X}{target data.}

\item{maxdimension}{max dimension.}

\item{maxscale}{max scale.}

\item{plot}{if \code{T}, plot diagram.}

\item{ret}{if \code{T}, return diagram.}
}
\value{
persistent diagram.
}
\description{
This function overwrites global variant \code{diagram} to prevent to lose calculateion result of persistent diagram.
}
\examples{

library(TDA)

circle <- circleUnif(100)

circle.diag <- calcPhom(circle, maxdimension = 1, maxscale = 1)

}
