% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_phom.R
\name{calcSubsamplePhom}
\alias{calcSubsamplePhom}
\title{Calculate persistent diagram of subsample of point cloud \code{X}}
\usage{
calcSubsamplePhom(X, maxdimension, maxscale, rate = 0.5, plot = T,
  ret = F)
}
\arguments{
\item{X}{target data.}

\item{maxdimension}{max dimension.}

\item{maxscale}{max scale.}

\item{rate}{sabsampling rate. larger rate takes large subsample.}

\item{plot}{if \code{T}, plot diagram.}

\item{ret}{if \code{T}, return diagram.}
}
\value{
persistent diagram.
}
\description{
Calculate persistent diagram of subsample of point cloud \code{X}
}
