% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_phom.R
\name{formatDiagram}
\alias{formatDiagram}
\title{Formats persistent diagram}
\usage{
formatDiagram(diag)
}
\arguments{
\item{diag}{persistent diagram.}
}
\value{
formatted persistent diagram.
}
\description{
Solve problem caused by difference of persistent diagram format between \code{myphom} and \code{TDA} package.
}
