% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_phom.R
\name{showPersistentDiagram}
\alias{showPersistentDiagram}
\title{Shows persistent diagram}
\usage{
showPersistentDiagram(diag = diagram, maxdimention = NULL,
  maxscale = NULL, cex = 1.5)
}
\arguments{
\item{diag}{persistent diagram.}

\item{maxdimention}{max dimension.}

\item{maxscale}{max scale.}

\item{cex}{graphics parameters.}
}
\description{
Shows persistent diagram
}
